Update usage notebook, script and bash files
